CustomYieldInstruction
