+++ Assets/Behaviours/EaserAlpha.cs
