+++ Assets/Behaviours/EaserColorAlpha.cs
