--- Assets/RealTime.cs.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: e672ae2a7208f95449f201e6603e6ace
-labels:
-- Code
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
