--- Assets/Singleton.cs.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 0d0c942fdb84c664bac142c4f6636796
-labels:
-- Code
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
