--- Assets/Utility.cs.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: e739fd7024e4e4a489061c0f329a65da
-labels:
-- Code
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
